pepare to marge
