+++ src/components/users/user.module.scss
@@ -1,0 +1,131 @@
+@font-face {
+  font-family: roboto;
+  src: url("./../../fonts/Roboto-Regular.ttf");
+}
+.room {
+  position: absolute;
+  width: 1366px;
+  height: 768px;
+  left: 0px;
+  top: 0px;
+  background: linear-gradient(
+    180deg,
+    rgba(0, 0, 0, 0) 0%,
+    rgba(0, 0, 0, 0.24) 100%
+  );
+}
+.roomcontainer {
+  position: absolute;
+  width: 1227px;
+  height: 653px;
+  left: 68px;
+  top: 54px;
+
+  background: #ffffff;
+  opacity: 0.45;
+}
+.roomname {
+  position: absolute;
+  width: 141px;
+  height: 43px;
+  left: 17px;
+  top: 0px;
+
+  font-family: roboto;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 14px;
+  line-height: 16px;
+  text-align: center;
+  letter-spacing: 0.0357143em;
+
+  color: #000000;
+}
+.logodown {
+  position: absolute;
+  width: 141px;
+  height: 99px;
+  left: 19px;
+  top: 559px;
+
+  background: url("./../../images/snackbla4.png");
+  border-radius: 30px;
+}
+.mic {
+  position: absolute;
+  width: 51px;
+  height: 51px;
+  left: 427px;
+  top: 545px;
+
+  background: url("./../../images/mic.png");
+}
+.arrow {
+  position: absolute;
+  width: 51px;
+  height: 51px;
+  left: 753px;
+  top: 545px;
+
+  background: url("./../../images/arrow.png");
+}
+.userslist {
+  position: absolute;
+  width: 51px;
+  height: 51px;
+  left: 1169px;
+  top: 9px;
+
+  background: url("./../../images/users.png");
+}
+.user1 {
+  position: absolute;
+  width: 328px;
+  height: 212px;
+  left: 194px;
+  top: 41px;
+  background-color: #ac7d7d;
+}
+.user2 {
+  position: absolute;
+  width: 328px;
+  height: 212px;
+  left: 716px;
+  top: 41px;
+  background-color: #927474;
+}
+.user3 {
+  position: absolute;
+  width: 328px;
+  height: 212px;
+  left: 455px;
+  top: 314px;
+  background-color: #8d4747;
+}
+
+.usernamecontainer {
+  position: absolute;
+  width: 109px;
+  height: 31px;
+  left: 0px;
+  top: 181px;
+
+  background: #000000;
+}
+.username {
+  position: absolute;
+  width: 42px;
+  height: 16px;
+  left: 33px;
+  top: -5px;
+
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 14px;
+  line-height: 16px;
+  text-align: center;
+  letter-spacing: 0.0357143em;
+
+  color: #ffffff;
+}
